requirements.txt
pyTelegramBotAPI==4.12.0
gspread==5.12.0
oauth2client==4.1.3
requests>=2.31.0
google-auth>=2.29.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.10.0